Split the v2.1 and the v5 integration tests
