+++ timbuctoo-instancev4/src/test/resources/nl/knaw/huygens/timbuctoo/v5/2017_04_17_BIA_Clusius.xlsx
